Move levels into Resources folder
So Unity is really dumb and doesn't actually build text files.
Apparently the Resources folder is where we're supposed to store
any file that we need to access from the code. And it has to be called
Resources, and the extension isn't specified when you load the file,
because REASONS.