Implement pawn captures
== begin C# rant ==

I searched "C# any" and discovered the "Any" method on Enumerable. Okay,
this makes sense. This is a method to determine whether any list element
satisfies a prediate, just like in every other language ever, right?

WRONG.

The "Any" method checks whether a list contains any elements. As in,
whether the length is not zero.

WHY?!?!?!

First of all, this is essentially equivalent to !list.Empty(), which is
the standard method for almost every single method. And even if we don't
consider this to be one of the most extremely useless methods that could
be created...

The /actual/ method name for checking if any item in a list matches a
predicate...

... is...

... Exists.

...

I have no words.

== end rant ==