SCROLLING WORKS
Time for... ANOTHER UNITY RANT! This one's shorter, but no less
infuriating.

First of all, apparently the correct way to do scrolling is to nest the
elements being scrolled FOUR layers deep. Canvas -> ScrollRect container ->
Viewport container -> Content container -> ACTUAL THINGS. This is super
dumb.

Secondly, the effect that the size and position of the transforms and
RectTransforms of ALL FOUR of these elements have on the actual outcome
is wildly unpredictable and not documented anywhere, as far as I can
tell.

Wow, I really do love Unity.