Make it look less terrible
You would think that I'd run out of things to rant about by now...

I added spacing between the buttons in the scroll view because, you
know, I wanted it to actually look nice. Only to discover that it would
randomly stop scrolling.

Turns out, you have to be hovering the mouse OVER THE CONTENT in order
for the scrollview to scroll. Not the viewport, not the content
container, but the ACTUAL CONTENT ELEMENTS.

So while I was scrolling through, if my mouse happened to hit an empty
space, it would STOP.

And I would have to move the mouse back onto a button in order to start
scrolling again.

I started to add a vertical scrollbar, but then I ran out of soul for
Unity to suck out.